pyproject.toml
[project]
name = "epl-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dbt-core>=1.11.7",
    "dbt-snowflake>=1.11.3",
    "dotenv>=0.9.9",
    "kaggle>=2.0.0",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "pyspark>=4.1.1",
    "snowflake-connector-python>=4.4.0",
]
